Entire chain of MainKKMC works correctly
git-svn-id: svn+ssh://fwe02.ifj.edu.pl/xdat1/lhcb/mchrzaszcz/svn/KK-all-2012@682 957d80b3-de0b-4737-8e94-343ebb428fb4

+++ MaMar/Plots/KKplot.h
@@ -34,25 +34,25 @@ using namespace std;
 
 //------------------------------------------------------------
 //  wrappers to f77 routines in KKMC and KKsem
-extern "C" void kk2f_fort_open_( const long&, char*, long);
-extern "C" void kk2f_fort_close_(const long&);
+extern "C" void kk2f_fort_open_( const int&, char*, int);
+extern "C" void kk2f_fort_close_(const int&);
 //      SUBROUTINE KKsem_Initialize(xpar_input)
 extern "C" void kksem_initialize_(double xpar[]);
 //      SUBROUTINE KK2f_Initialize(xpar)
 extern "C" void kk2f_initialize_(double xpar[]);
 //-----------------------
 //      SUBROUTINE KKsem_SetKFfin(KFfin)
-extern "C" void kksem_setkffin_( const long& );
+extern "C" void kksem_setkffin_( const int& );
 //      SUBROUTINE BornV_SetKF(KFferm)
-extern "C" void bornv_setkf_( const long& ); // set SINGLE Final State
+extern "C" void bornv_setkf_( const int& ); // set SINGLE Final State
 //      SUBROUTINE KKsem_SetKeyFoB(KeyFoB)
-extern "C" void kksem_setkeyfob_( const long& );
+extern "C" void kksem_setkeyfob_( const int& );
 //      SUBROUTINE KKsem_VVplot_vec(key,chak,nbin,xmin,xmax,yy)
-extern "C" void kksem_vvplot_vec_(const long&, char[5], const long&, const double&, const double&, double[]);
+extern "C" void kksem_vvplot_vec_(const int&, char[5], const int&, const double&, const double&, double[]);
 //      SUBROUTINE KKsem_SetCrange(Cmin,Cmax)
 extern "C" void kksem_setcrange_(const double&, const double&);
 //      SUBROUTINE KKsem_SetKeyZet(KeyZet)
-extern "C" void kksem_setkeyzet_( const long& );
+extern "C" void kksem_setkeyzet_( const int& );
 //      SUBROUTINE KKsem_MakeBorn(svar,Born)
 extern "C" void kksem_makeborn_(const double&, double&);
 //---------------------------------
@@ -61,21 +61,21 @@ extern "C" double bornv_sig0nb_(const double&);
 //      SUBROUTINE BornV_MakeGami(CMSene,gamiCR,gami,alfi)
 extern "C" void bornv_makegami_(const double&, double&, double&, double&);
 //      DOUBLE PRECISION  FUNCTION BornV_Simple(KFi,KFf,svar,costhe)
-extern "C" double bornv_simple_( const long&,  const long&, const double&, const double&);
+extern "C" double bornv_simple_( const int&,  const int&, const double&, const double&);
 //------------------------------------------------------------
 //      SUBROUTINE BornV_InterpoGSW(KFf,svar,CosThe)
-extern "C" double bornv_interpogsw_( const long&,  const double&, const double&);
+extern "C" double bornv_interpogsw_( const int&,  const double&, const double&);
 //      DOUBLE PRECISION FUNCTION BornV_Dizet(Mode,KFi,KFf,svar,CosThe,eps1,eps2,ta,tb)
-extern "C" double bornv_dizet_(const long&, const long&, const long&,
+extern "C" double bornv_dizet_(const int&, const int&, const int&,
 		const double&, const double&, const double&, const double&, const double&, const double& );
 //------------------------------------------------------------
 //      SUBROUTINE GPS_BornF(KFi,KFf,PX,CosThe,p1,m1,p2,m2,p3,m3,p4,m4,Xborn)
-extern "C" void gps_bornf_(const long&, const long&, double[], const double&,
+extern "C" void gps_bornf_(const int&, const int&, double[], const double&,
 		    double[], const double&, double[], const double&, double[], const double&, double[], const double&,
 		    const double&);
 //------------------------------------------------------------
 //      SUBROUTINE GPS_BornFoam(Mode,KFi,KFf,CMSene,CosThe,Xborn)
-extern "C" void gps_bornfoam_(const long&,   const long&,   const long&,
+extern "C" void gps_bornfoam_(const int&,   const int&,   const int&,
 		                      const double&, const double&, const double&);
 //      DOUBLE PRECISION  FUNCTION GPS_MakeRhoFoam(XNorm)
 extern "C" double gps_makerhofoam_(const double&);
@@ -103,8 +103,8 @@ class KKplot{
  	//
  	double m_fin;
  	double m_chfin;
- 	long   m_KFini;  // electron
- 	long   m_KFf;    // muon
+ 	int   m_KFini;  // electron
+ 	int   m_KFf;    // muon
  	//
  	int    m_kDim;
  	int    m_nCells;
@@ -121,8 +121,8 @@ class KKplot{
 public:
   // Interfaces to KKsem integration routines using Gauss method
   void Initialize(TFile&);
-  void VVplot( TH1 *, long , char [], long, long );
-  void Cplot(  TH1 *, long , char [], long, long, double, double);
+  void VVplot( TH1 *, int , char [], int, int );
+  void Cplot(  TH1 *, int , char [], int, int, double, double);
 
 ////////////////////////////////////////////////////////////////////////////
 };// KKplot
